pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-tales"
version = "0.1.0"
description = "Transform GitHub repositories into narrated audio stories using Claude AI and ElevenLabs TTS"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "krzemienski" }]
requires-python = ">=3.10"
keywords = ["ai", "audio", "tts", "github", "storytelling", "claude", "elevenlabs"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "anthropic>=0.40.0",
    "httpx>=0.25.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
tts = ["elevenlabs>=1.0"]
dev = ["pytest>=7.0", "black", "ruff", "mypy"]

[project.scripts]
code-tales = "code_tales.cli:cli"

[project.urls]
Homepage = "https://github.com/krzemienski/code-tales"
Repository = "https://github.com/krzemienski/code-tales"
Issues = "https://github.com/krzemienski/code-tales/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/code_tales"]

[tool.hatch.build.targets.sdist]
include = ["src/", "README.md", "LICENSE", "pyproject.toml"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
